backend/requirements.txt
# FastAPI Backend Dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
sqlalchemy==2.0.25
databases[aiosqlite]==0.8.0
aiosqlite==0.19.0
pydantic==2.5.3
pydantic-settings==2.1.0
websockets==12.0
python-dotenv==1.0.0

# Existing bot dependencies
aiohttp==3.9.1
py-clob-client==0.20.0
python-dateutil==2.8.2
